Fix incorrect key extraction in InnerJoinOperation
The inner join on more than two datasets is implemented by resursively
joining every dataset after the second with the previous join. Such as

t1 \__..._ result
t2 / /  /
t3 _/  /
...   /
tn __/

The first dataset normalize the DataPoint to the final structure to make
each merge operation easy, but the first data structure was incorrectly
assumed to be the same (that is, the identifier of t1 and t2 to be at
the same place). This worked because the serialization in the connector
orders the column with the identifiers first, but this assumption is
wrong.